pyproject.toml
[project]
name = "neuro-cloud-api"
version = "1.0.0"
description = "Python библиотека для работы с облачными хранилищами (Яндекс.Диск, Google Drive, S3)"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "DOCUMENTATION.md"
license = {text = "LICENSE"}
requires-python = ">=3.9,<4.0"
dependencies = [
    "yadisk>=3.4.0",
    "python-dotenv>=1.0.0",
]

[tool.poetry]
name = "neuro-cloud-api"
version = "1.0.0"
description = "Python библиотека для работы с облачными хранилищами (Яндекс.Диск, Google Drive, S3)"
authors = ["Your Name <your.email@example.com>"]
readme = "DOCUMENTATION.md"
license = "LICENSE"
packages = [{include = "neuro_cloud_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
yadisk = "^3.4.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
